refined api endpoint for performance optimization
